Incorporate HW 4 feedback
Q1) "7/10 The size of two heap might become unbalanced during the insert
operation." It's now stressed that INSERT always maintains the balance
in two heaps.

Q6) "10/16 Good approach, but the second part of your algorithm is not
solid, or may run longer. Paths you already choose may change if setting
another edge to zero." A more fundamental problem is that I don't know
whether it's correct or not. Changed to the algorithm given in the
solution.